IH etc astype(float)

+++ chilbolton_ts_utils.py
@@ -1709,11 +1709,11 @@ def convert_galileo_ts_l0a2l1(infile,outfile,dBZ_offset,range_offset,data_versio
     print(txpol);
 
 
-    IH = DSin['IH'][firstray:,:,:];
-    QH = DSin['QH'][firstray:,:,:];
+    IH = DSin['IH'][firstray:,:,:].astype(float);
+    QH = DSin['QH'][firstray:,:,:].astype(float);
 
-    IV = DSin['IV'][firstray:,:,:];
-    QV = DSin['QV'][firstray:,:,:];
+    IV = DSin['IV'][firstray:,:,:].astype(float);
+    QV = DSin['QV'][firstray:,:,:].astype(float);
 
     nray     = IH.shape[0];
     npulse   = IH.shape[1];
@@ -1729,7 +1729,7 @@ def convert_galileo_ts_l0a2l1(infile,outfile,dBZ_offset,range_offset,data_versio
     QH[:,:,:] = QH[:,:,:]-QH_mean[:,None,:];
 
     print(IH);
-    
+
     Zcal=dBZ_offset+10*np.log10(DSout['prf'][:]/512);
 
     IH=IH*np.sqrt(10**(Zcal/10.0));
